+++ src/FFT.cpp
@@ -402,7 +402,7 @@ void NewWindowFunc(int whichFunction, size_t NumSamples, bool extraSample, float
       // Hamming
       const double multiplier = 2 * M_PI / NumSamples;
       static const double coeff0 = 0.54, coeff1 = -0.46;
-      for (int ii = 0; ii < NumSamples; ++ii)
+      for (size_t ii = 0; ii < NumSamples; ++ii)
          in[ii] *= coeff0 + coeff1 * cos(ii * multiplier);
    }
       break;
@@ -411,7 +411,7 @@ void NewWindowFunc(int whichFunction, size_t NumSamples, bool extraSample, float
       // Hanning
       const double multiplier = 2 * M_PI / NumSamples;
       static const double coeff0 = 0.5, coeff1 = -0.5;
-      for (int ii = 0; ii < NumSamples; ++ii)
+      for (size_t ii = 0; ii < NumSamples; ++ii)
          in[ii] *= coeff0 + coeff1 * cos(ii * multiplier);
    }
       break;
@@ -421,7 +421,7 @@ void NewWindowFunc(int whichFunction, size_t NumSamples, bool extraSample, float
       const double multiplier = 2 * M_PI / NumSamples;
       const double multiplier2 = 2 * multiplier;
       static const double coeff0 = 0.42, coeff1 = -0.5, coeff2 = 0.08;
-      for (int ii = 0; ii < NumSamples; ++ii)
+      for (size_t ii = 0; ii < NumSamples; ++ii)
          in[ii] *= coeff0 + coeff1 * cos(ii * multiplier) + coeff2 * cos(ii * multiplier2);
    }
       break;
@@ -432,7 +432,7 @@ void NewWindowFunc(int whichFunction, size_t NumSamples, bool extraSample, float
       const double multiplier2 = 2 * multiplier;
       const double multiplier3 = 3 * multiplier;
       static const double coeff0 = 0.35875, coeff1 = -0.48829, coeff2 = 0.14128, coeff3 = -0.01168;
-      for (int ii = 0; ii < NumSamples; ++ii)
+      for (size_t ii = 0; ii < NumSamples; ++ii)
          in[ii] *= coeff0 + coeff1 * cos(ii * multiplier) + coeff2 * cos(ii * multiplier2) + coeff3 * cos(ii * multiplier3);
    }
       break;
@@ -440,7 +440,7 @@ void NewWindowFunc(int whichFunction, size_t NumSamples, bool extraSample, float
    {
       // Welch
       const float N = NumSamples;
-      for (int ii = 0; ii < NumSamples; ++ii) {
+      for (size_t ii = 0; ii < NumSamples; ++ii) {
          const float iOverN = ii / N;
          in[ii] *= 4 * iOverN * (1 - iOverN);
       }
@@ -452,7 +452,7 @@ void NewWindowFunc(int whichFunction, size_t NumSamples, bool extraSample, float
       // Precalculate some values, and simplify the fmla to try and reduce overhead
       static const double A = -2 * 2.5*2.5;
       const float N = NumSamples;
-      for (int ii = 0; ii < NumSamples; ++ii) {
+      for (size_t ii = 0; ii < NumSamples; ++ii) {
          const float iOverN = ii / N;
          // full
          // in[ii] *= exp(-0.5*(A*((ii-NumSamples/2)/NumSamples/2))*(A*((ii-NumSamples/2)/NumSamples/2)));
@@ -466,7 +466,7 @@ void NewWindowFunc(int whichFunction, size_t NumSamples, bool extraSample, float
       // Gaussian (a=3.5)
       static const double A = -2 * 3.5*3.5;
       const float N = NumSamples;
-      for (int ii = 0; ii < NumSamples; ++ii) {
+      for (size_t ii = 0; ii < NumSamples; ++ii) {
          const float iOverN = ii / N;
          in[ii] *= exp(A * (0.25 + (iOverN * iOverN) - iOverN));
       }
@@ -477,7 +477,7 @@ void NewWindowFunc(int whichFunction, size_t NumSamples, bool extraSample, float
       // Gaussian (a=4.5)
       static const double A = -2 * 4.5*4.5;
       const float N = NumSamples;
-      for (int ii = 0; ii < NumSamples; ++ii) {
+      for (size_t ii = 0; ii < NumSamples; ++ii) {
          const float iOverN = ii / N;
          in[ii] *= exp(A * (0.25 + (iOverN * iOverN) - iOverN));
       }
@@ -540,7 +540,7 @@ void DerivativeOfWindowFunc(int whichFunction, size_t NumSamples, bool extraSamp
       wxASSERT(NumSamples > 0);
       --NumSamples;
       // in[0] *= 1.0f;
-      for (int ii = 1; ii < NumSamples; ++ii)
+      for (size_t ii = 1; ii < NumSamples; ++ii)
          in[ii] = 0.0f;
       in[NumSamples] *= -1.0f;
       return;
@@ -592,7 +592,7 @@ void DerivativeOfWindowFunc(int whichFunction, size_t NumSamples, bool extraSamp
       in[0] *= coeff0;
       if (!extraSample)
          --NumSamples;
-      for (int ii = 0; ii < NumSamples; ++ii)
+      for (size_t ii = 0; ii < NumSamples; ++ii)
          in[ii] *= - coeff1 * sin(ii * multiplier);
       if (extraSample)
          in[NumSamples] *= - coeff0;
@@ -606,7 +606,7 @@ void DerivativeOfWindowFunc(int whichFunction, size_t NumSamples, bool extraSamp
       // Hanning
       const double multiplier = 2 * M_PI / NumSamples;
       const double coeff1 = -0.5 * multiplier;
-      for (int ii = 0; ii < NumSamples; ++ii)
+      for (size_t ii = 0; ii < NumSamples; ++ii)
          in[ii] *= - coeff1 * sin(ii * multiplier);
       if (extraSample)
          in[NumSamples] = 0.0f;
@@ -618,7 +618,7 @@ void DerivativeOfWindowFunc(int whichFunction, size_t NumSamples, bool extraSamp
       const double multiplier = 2 * M_PI / NumSamples;
       const double multiplier2 = 2 * multiplier;
       const double coeff1 = -0.5 * multiplier, coeff2 = 0.08 * multiplier2;
-      for (int ii = 0; ii < NumSamples; ++ii)
+      for (size_t ii = 0; ii < NumSamples; ++ii)
          in[ii] *= - coeff1 * sin(ii * multiplier) - coeff2 * sin(ii * multiplier2);
       if (extraSample)
          in[NumSamples] = 0.0f;
@@ -632,7 +632,7 @@ void DerivativeOfWindowFunc(int whichFunction, size_t NumSamples, bool extraSamp
       const double multiplier3 = 3 * multiplier;
       const double coeff1 = -0.48829 * multiplier,
          coeff2 = 0.14128 * multiplier2, coeff3 = -0.01168 * multiplier3;
-      for (int ii = 0; ii < NumSamples; ++ii)
+      for (size_t ii = 0; ii < NumSamples; ++ii)
          in[ii] *= - coeff1 * sin(ii * multiplier) - coeff2 * sin(ii * multiplier2) - coeff3 * sin(ii * multiplier3);
       if (extraSample)
          in[NumSamples] = 0.0f;
@@ -643,7 +643,7 @@ void DerivativeOfWindowFunc(int whichFunction, size_t NumSamples, bool extraSamp
       // Welch
       const float N = NumSamples;
       const float NN = NumSamples * NumSamples;
-      for (int ii = 0; ii < NumSamples; ++ii) {
+      for (size_t ii = 0; ii < NumSamples; ++ii) {
          in[ii] *= 4 * (N - ii - ii) / NN;
       }
       if (extraSample)
@@ -675,7 +675,7 @@ void DerivativeOfWindowFunc(int whichFunction, size_t NumSamples, bool extraSamp
       in[0] *= exp(A * 0.25) * (1 - invN);
       if (!extraSample)
          --NumSamples;
-      for (int ii = 1; ii < NumSamples; ++ii) {
+      for (size_t ii = 1; ii < NumSamples; ++ii) {
          const float iOverN = ii * invN;
          in[ii] *= exp(A * (0.25 + (iOverN * iOverN) - iOverN)) * (2 * ii * invNN - invN);
       }
